Revert 'rename submodule `region_markers`->`markers`'
